[compiler] remove cat aggressive mode
