+++ compiler/include/byteir/Pipelines/HloFusionOpt.h
@@ -47,10 +47,6 @@ struct HloFusionOptPipelineOptions
       *this, "outline-cat-op",
       llvm::cl::desc("whether to outline cat ops and AIT as an backend"),
       llvm::cl::init(false)};
-  Option<bool> aggressiveCatFusion{
-      *this, "aggressive-cat-fusion",
-      llvm::cl::desc("whether to fuse CAT aggressively"),
-      llvm::cl::init(false)};
 };
 
 void createHloFusionOptPipeline(OpPassManager &pm,
