+++ tests/numerical_test/main.py
@@ -88,6 +88,7 @@ CUDA_AIT_MLIR_TEST_SET = {
     "gemm_rrr_f32.mlir",
     "layernorm.mlir",
     "softmax.mlir",
+    "transpace2d.mlir",
     "transpose102.mlir",
     "transpose1023.mlir",
     "transpose120.mlir",
@@ -179,7 +180,6 @@ def parse_args():
             "cpu",
             "cuda",
             "cuda_with_ait",
-            "cuda_with_ait_aggressive",
             "dynamo",
             "native_torch",
         ],
